Add testing for array, unmarshal and pubsub
Also contains a significant rewrite of array and some more usage of the
context variable in pubsub. There are some missing tests in pubsub still
and pubsub currently does not work after the rewrite of command to
ArrayWriter.  The hope is to address this as I finish writing the tests
for the pubsub components.

I do want to move to using more of the io and ioutil functions and
structs of the standard library when reading and writing to the
connection as reusing a single buffer for writing to the connection
turned out to be naive and the native implementations of bufio handles
writing to sockets much better and efficiently.