Change from (un)marshal use io.Reader features
This principally rips out the use of raw []byte buffers and replaces it
with a bufio.Reader for reading off of the connection. We still do size
down buffer sizes as before, it however only does it for the reader and
not for the writer. We may implement that later.