Test coverage and alloc management
Coverage is back up to 100% and all allocs for PubSub has been
identified and removed. There's only one major thing left and that is to
set up the benchmarks for the competing libraries and publish those in
the README and also update the other two libraries to the most recent
versions.